pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "airs"
version = "0.1.0"
description = "AI Research System — end-to-end autonomous research pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49",
    "openai>=1.60",
    "google-generativeai",
    "paramiko>=3.4",
    "typer>=0.12",
    "rich>=13.7",
    "fastapi>=0.111",
    "uvicorn>=0.30",
    "jinja2>=3.1",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "tenacity>=8.3",
    "structlog>=24.1",
    "arxiv>=2.1",
    "markdown>=3.6",
    "watchdog>=4.0",
    "httpx>=0.27",
    "duckduckgo-search>=6.0",
]

[project.scripts]
airs = "airs.main:app"

[tool.hatch.build.targets.wheel]
packages = ["airs"]
